implement persistence

+++ app/src/main/java/co/edu/intecap/notes/listeners/NoteEventListener.java
@@ -1,5 +1,5 @@
 package co.edu.intecap.notes.listeners;
 
 public interface NoteEventListener {
-    void onNoteSelected(int noteId);
+    void onNoteSelected(long noteId);
 }
